refactor: move constraint functinos to separate file
